refactor(diffusion): Convert module to package structure
Split diffusion.py (460 lines) into focused submodules:
- core.py: Main Diffusion class orchestrator
- validators.py: Input validation (validate_repo_identifier, validate_credential_type)
- resolvers.py: Identifier resolution (resolve_shortname_to_id, resolve_object_identifier)
- fetchers.py: API data fetching (fetch_repositories, fetch_branches, fetch_uris)
- formatters.py: Data formatting (format_uris, format_repositories, format_branches)

Maintains backward compatibility via __init__.py re-exports.
Similar refactoring pattern as maniphest package from PR #126.
